organizing files into respective folders
