+++ childPiece/bishop.py
@@ -1,7 +1,7 @@
-from piece import Piece
+from parentPiece.piece import Piece
 
 
-class Queen(Piece):
+class Bishop(Piece):
     def __init__(self, pMove, pPoints, pColor):
         self.move = pMove
         self.points = pPoints
