+++ childPiece/king.py
@@ -1,7 +1,7 @@
-from piece import Piece
+from parentPiece.piece import Piece
 
 
-class Rook(Piece):
+class King(Piece):
     def __init__(self, pMove, pPoints, pColor):
         self.move = pMove
         self.points = pPoints
