+++ childPiece/rook.py
@@ -1,7 +1,7 @@
-from piece import Piece
+from parentPiece.piece import Piece
 
 
-class Knight(Piece):
+class Rook(Piece):
     def __init__(self, pMove, pPoints, pColor):
         self.move = pMove
         self.points = pPoints
