--- readme.txt
@@ -1,28 +1,0 @@
-2 players on one computer. No need localhost or internet connectivity
-In Python 3.8.3, GitHub hosted by Mariam, git-bash (or similar), userstories in some software and Pycharm
-
-ui with Tim and Connor:
-	PyGame or SimplePyGui (Probably PyGame)
-	Thumbnails of the pieces
-	chessboard
-	each piece linking up with the corresponding class in the backend
-	scoreboard
-	On first click, show where piece can move. Then move the piece or put it back down.
-
-api with Peter and Mariam:
-	Mostly logic
-	Classes of each piece
-		Parent called "Piece"
-		Child classes of piece
-	Win condition
-		Some way to tell if someone has won or not
-	Scoring with player 1 and player 2
-	Try to unittest
-
-Onboarding + goals:
-	Getting used to user stories and asigning points and git shtuff
-		Base branch of develop -> branch off of it for each user story -> code -> Do a pull request (Tim reviewing)
-		Don't merge with develop ever
-		Develop is the shared of everyone
-	Working as a team
-	Try not to look "how to make chess app in python" or something similar
